tmtc-c2a: use c2a-devtools-frontend from cargo deps
