+++ tmtc-c2a/build.rs
@@ -4,25 +4,6 @@ use std::{
     path::{Path, PathBuf},
 };
 
-fn wasm_packages_root() -> PathBuf {
-    let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
-    out_dir.join("wasm_packages")
-}
-
-fn wasm_pack(name: &str, devtools_build_dir: &PathBuf) {
-    let pkg_outdir = wasm_packages_root().join(name).join("pkg");
-    let status = Command::new("pnpm")
-        .current_dir(devtools_build_dir)
-        .arg("run")
-        .arg("crate")
-        .arg(name)
-        .arg("--out-dir")
-        .arg(&pkg_outdir)
-        .status()
-        .expect("failed to build frontend");
-    assert!(status.success());
-}
-
 fn main() {
     let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
 
@@ -31,59 +12,5 @@ fn main() {
         .compile(&["./proto/tmtc_generic_c2a.proto"], &["./proto"])
         .unwrap_or_else(|e| panic!("Failed to compile protos {:?}", e));
 
-    if std::env::var("SKIP_FRONTEND_BUILD").is_err() {
-        println!("cargo:rerun-if-changed=devtools_frontend");
-
-        // copy frontend source into OUT_DIR
-        let devtools_build_dir = out_dir.join("devtools_frontend");
-        copy_devtools_dir("devtools_frontend", &devtools_build_dir).unwrap();
-
-        let status = Command::new("corepack")
-            .arg("enable")
-            .current_dir(&devtools_build_dir)
-            .status()
-            .expect("failed to execute corepack");
-        assert!(status.success(), "failed to install pnpm via corepack");
-
-        let status = Command::new("pnpm")
-            .arg("install")
-            .current_dir(&devtools_build_dir)
-            .status()
-            .expect("failed to execute pnpm");
-        assert!(status.success(), "failed to install deps for frontend");
-
-        wasm_pack("wasm-opslang", &devtools_build_dir);
-
-        let devtools_out_dir = out_dir.join("devtools_dist");
-        let status = Command::new("pnpm")
-            .current_dir(&devtools_build_dir)
-            // vite.config.ts にwasmのビルド場所を教えるために環境変数を渡す
-            .envs([("DEVTOOLS_CRATE_ROOT", wasm_packages_root())])
-            .arg("run")
-            .arg("build:vite")
-            .arg("--outDir")
-            .arg(&devtools_out_dir)
-            .status()
-            .expect("failed to execute pnpm");
-        assert!(status.success(), "failed to build frontend");
-    }
-
     notalawyer_build::build();
-}
-
-fn copy_devtools_dir(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> io::Result<()> {
-    fs::create_dir_all(&dst)?;
-    for entry in fs::read_dir(src)? {
-        let entry = entry?;
-        let ty = entry.file_type()?;
-        if ty.is_dir() {
-            if entry.file_name().to_str() == Some("node_modules") {
-                continue;
-            }
-            copy_devtools_dir(entry.path(), dst.as_ref().join(entry.file_name()))?;
-        } else {
-            fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
-        }
-    }
-    Ok(())
 }
